Avoid deprecation warning
"DeprecationWarning: The truth value of an empty array is ambiguous.
Returning False, but in future this will result in an error.
Use array.size > 0` to check that an array is not empty."

We want to retain flexibility with regard to types of sequences, so
convert all to np.array.  In most cases, this should be a null op.