Raise minimum nexgen version to support older data
Some early Tristan data recorded before March 2021, and where the
goniometer was not moved during the data collection, record the rotation
axis position as a scalar, whereas we expect it always to be a tuple.
Support for this legacy data format in Nexgen was introduced in
dials/nexgen#8, which was released in Nexgen version 0.4.8.