Revert "Further simplify the task graph"
We cannot do a single bincount because that requires having enough
memory for the result, which is the entire image stack.  This defeats
the object of using Dask in the first place, to enable working with
larger-than-memory data sets.  This isn't the first time I've fallen
into this trap, so add a comment this time.

This reverts commit f971be974726425e5a9508ef031f5bd649361ebf.