Avoid creating excess chunks for slicing
When slicing an array, and specifying the index with np.flatnonzero in
place of a boolean array, to allow for pre-computed chunk sizes of the
index array in cases where the index array is reused, avoid comparing
every chunk of the index array with every chunk of the array to be
sliced.  Instead, simply perform the slicing as a pairwise operation,
with each chunk of the array to be sliced paired with the corresponding
chunk from the index array.