Select block-wise, don't compute_chunk_sizes
Dask slicing returns an array with unknown chunk sizes, which causes
problems for downstream Dask operations.  Given that we are simply
lumping the resulting arrays into a bincount call, and have no need of
the chunk size information, we can save ourselves an expensive pass over
the data by not computing the chunk sizes at all.  This can be achieved
by replacing Dask slicing with NumPy slicing, mapped over all the blocks
of an array.