Refactor to better manage scheduler memory load (#89)
This refactor makes two significant alterations.
  -  The basic data structure for reading raw LATRD data is changed from a dataclass with Dask array field values to a Dask dataframe, allowing datasets that have a natural association with one another to be sliced together intuitively.
  - The mechanism for binning images is changed to allow the scheduler to drop references to intermediate results more promptly, and to improve the size of the task graph and the parallelisation of tasks.

Before this change set, the task graph for binning events to images included a few layers of data sanitisation, one branch of tasks per input chunk, and a layer of binning tasks, one bincount task per image per input chunk.  The task graph scaled as the product of the number of input chunks and the number of images, which could cause the task graph construction to be prohibitively costly.  Additionally, each of the bincount tasks depended on the results of all the input/sanitisation branches, meaning that the scheduler would try to keep references to all the sanitisation results until the last bincount was complete.  The memory load was often unacceptable.

Now, the image binning layer consists of a single task acting on the results of each of the data sanitisation tasks, therefore simply scaling with the number of input chunks.  The binning task adds the resulting partial images to an on-disk image stack cache, consisting of a Zarr array.  To avoid a race condidtion when different tasks are trying to add to the same image simultaneously, the binning is performed using a dask.distributed.Lock.

The change to using Dask dataframe objects to handle the input data has necessitated various small changes to tools that previously handled Dask arrays.