Simplify progress bar for distributed computation
Display the progress of the full task graph, not just the top layer.
Constraining ourselves to just the top layer should not be necessary if
we can serialise the file reading tasks approriately, which leads to
more fluid progress through the task graph.  Also, add option to gather
and return the computed futures.